+++ src/scorer/scorer.ts
@@ -44,12 +44,16 @@ export default abstract class Scorer {
 
     // This is the public API for scoring a toot
     async score(toot: Toot): Promise<number> {
-        if (!this.isReady) {
-            console.warn(`${this.name} not ready, scoring 0...`);
-            return 0;
-        }
+        if (this.isReady) return await this._score(toot);
 
-        return await this._score(toot);
+        if (!toot.scoreInfo) {
+            console.warn(`${this.logPrefix()} not ready, scoring 0...`);
+            return 0;
+        } else {
+            const existingScore = toot.scoreInfo.rawScores[this.name];
+            console.debug(`${this.logPrefix()} Not ready but toot already scored (existing score: ${existingScore})`);
+            return existingScore;
+        }
     }
 
     // Actual implementation of the scoring algorithm should be implemented in subclasses